+++ test/hw_tests/hwtest_zoneset_switching.cpp
@@ -28,6 +28,7 @@
 #include "psen_scan_v2_standalone/util/async_barrier.h"
 
 #include "psen_scan_v2/ros_integrationtest_helper.h"
+#include "psen_scan_v2/subscriber_mock.h"
 
 namespace psen_scan_v2_test
 {
