+++ test/integration_tests/integrationtest_active_zoneset_node.cpp
@@ -29,6 +29,8 @@
 #include <std_msgs/UInt8.h>
 
 #include "psen_scan_v2/ros_integrationtest_helper.h"
+#include "psen_scan_v2/subscriber_mock.h"
+
 #include "psen_scan_v2_standalone/util/async_barrier.h"
 #include "psen_scan_v2_standalone/util/gmock_expectations.h"
